added the service to access different words.
